+++ src/Oro/Bundle/ProductBundle/Tests/Unit/Provider/ProductImageFileNameProviderTest.php
@@ -6,6 +6,7 @@ use Oro\Bundle\AttachmentBundle\Entity\File;
 use Oro\Bundle\AttachmentBundle\Provider\FileNameProviderInterface;
 use Oro\Bundle\CatalogBundle\Entity\Category;
 use Oro\Bundle\ConfigBundle\Config\ConfigManager;
+use Oro\Bundle\FeatureToggleBundle\Checker\FeatureChecker;
 use Oro\Bundle\ProductBundle\Entity\ProductImage;
 use Oro\Bundle\ProductBundle\Provider\ProductImageFileNameProvider;
 
@@ -14,24 +15,25 @@ class ProductImageFileNameProviderTest extends \PHPUnit\Framework\TestCase
     /** @var FileNameProviderInterface|\PHPUnit\Framework\MockObject\MockObject */
     private $innerProvider;
 
-    /** @var ConfigManager|\PHPUnit\Framework\MockObject\MockObject */
-    private $configManager;
+    /** @var FeatureChecker|\PHPUnit\Framework\MockObject\MockObject */
+    private $featureChecker;
 
-    /** @var ProductImageFileNameProvider */
-    private $provider;
+    private ProductImageFileNameProvider $provider;
 
     protected function setUp(): void
     {
         $this->innerProvider = $this->createMock(FileNameProviderInterface::class);
-        $this->configManager = $this->createMock(ConfigManager::class);
+        $this->featureChecker = $this->createMock(FeatureChecker::class);
 
         $this->provider = new ProductImageFileNameProvider(
             $this->innerProvider,
-            $this->configManager
+            $this->createMock(ConfigManager::class)
         );
+        $this->provider->setFeatureChecker($this->featureChecker);
+        $this->provider->addFeature('product_original_filenames');
     }
 
-    public function testGetFileNameNotProductImage()
+    public function testGetFileNameNotProductImage(): void
     {
         $file = new File();
         $file->setFilename('filename.jpeg');
@@ -39,18 +41,21 @@ class ProductImageFileNameProviderTest extends \PHPUnit\Framework\TestCase
         $file->setExtension('jpeg');
         $file->setParentEntityClass(Category::class);
 
-        $this->configManager->expects($this->never())
-            ->method($this->anything());
+        $this->featureChecker->expects(self::never())
+            ->method(self::anything());
 
-        $this->innerProvider->expects($this->once())
+        $this->innerProvider->expects(self::once())
             ->method('getFileName')
             ->with($file)
             ->willReturn('filename.jpeg');
 
-        $this->assertEquals('filename.jpeg', $this->provider->getFileName($file));
+        self::assertEquals(
+            'filename.jpeg',
+            $this->provider->getFileName($file)
+        );
     }
 
-    public function testGetFileNameOriginalFileNamesDisabled()
+    public function testGetFileNameOriginalFileNamesDisabled(): void
     {
         $file = new File();
         $file->setFilename('filename.jpeg');
@@ -58,54 +63,117 @@ class ProductImageFileNameProviderTest extends \PHPUnit\Framework\TestCase
         $file->setExtension('jpeg');
         $file->setParentEntityClass(ProductImage::class);
 
-        $this->configManager->expects($this->once())
-            ->method('get')
-            ->with('oro_product.original_file_names_enabled')
+        $this->featureChecker->expects(self::once())
+            ->method('isFeatureEnabled')
+            ->with('product_original_filenames')
             ->willReturn(false);
 
-        $this->innerProvider->expects($this->once())
+        $this->innerProvider->expects(self::once())
             ->method('getFileName')
             ->with($file)
             ->willReturn('filename.jpeg');
 
-        $this->assertEquals('filename.jpeg', $this->provider->getFileName($file));
+        self::assertEquals(
+            'filename.jpeg',
+            $this->provider->getFileName($file)
+        );
     }
 
-    public function testGetFileNameOriginalFileNamesEnabled()
+    public function testGetFileNameFilenameSameAsOriginal(): void
+    {
+        $fileName = 'original-filename_#123-картинка))).jpeg';
+
+        $file = new File();
+        $file->setFilename($fileName);
+        $file->setOriginalFilename($fileName);
+        $file->setParentEntityClass(ProductImage::class);
+
+        $this->featureChecker->expects(self::once())
+            ->method('isFeatureEnabled')
+            ->with('product_original_filenames')
+            ->willReturn(true);
+
+        $this->innerProvider->expects(self::never())
+            ->method(self::anything());
+
+        self::assertEquals(
+            'original-filename_-123-картинка.jpeg',
+            $this->provider->getFileName($file)
+        );
+    }
+
+    public function testGetFileNameOriginalFileNamesEnabled(): void
     {
         $file = new File();
         $file->setFilename('filename.jpeg');
         $file->setOriginalFilename('original-filename_#123-картинка))).jpeg');
-        $file->setExtension('jpeg');
         $file->setParentEntityClass(ProductImage::class);
 
-        $this->configManager->expects($this->once())
-            ->method('get')
-            ->with('oro_product.original_file_names_enabled')
+        $this->featureChecker->expects(self::once())
+            ->method('isFeatureEnabled')
+            ->with('product_original_filenames')
             ->willReturn(true);
 
-        $this->innerProvider->expects($this->never())
-            ->method($this->anything());
+        $this->innerProvider->expects(self::never())
+            ->method(self::anything());
 
-        $this->assertEquals('filename-original-filename_-123-картинка.jpeg', $this->provider->getFileName($file));
+        self::assertEquals(
+            'filename-original-filename_-123-картинка.jpeg',
+            $this->provider->getFileName($file)
+        );
     }
 
-    public function testGetFileNameOriginalFileNamesEnabledNoOriginalFileName()
+    public function testGetFileNameOriginalFileNamesEnabledNoOriginalFileName(): void
     {
         $file = new File();
         $file->setFilename('filename.jpeg');
-        $file->setExtension('jpeg');
         $file->setParentEntityClass(ProductImage::class);
 
-        $this->configManager->expects($this->never())
-            ->method('get')
-            ->with('oro_product.original_file_names_enabled');
+        $this->featureChecker->expects(self::never())
+            ->method(self::anything());
 
-        $this->innerProvider->expects($this->once())
+        $this->innerProvider->expects(self::once())
             ->method('getFileName')
             ->with($file)
             ->willReturn('filename.jpeg');
 
-        $this->assertEquals('filename.jpeg', $this->provider->getFileName($file));
+        self::assertEquals(
+            'filename.jpeg',
+            $this->provider->getFileName($file)
+        );
+    }
+
+    /**
+     * @dataProvider getExtensionDataProvider
+     */
+    public function testGetFileName(?string $extension): void
+    {
+        $file = new File();
+        $file->setParentEntityClass(ProductImage::class);
+        $file->setFilename('filename.jpeg');
+        $file->setOriginalFilename('original-filename_#123-картинка))).jpeg');
+        $file->setExtension($extension);
+
+        $this->innerProvider->expects(self::never())
+            ->method(self::anything());
+
+        $this->featureChecker->expects(self::once())
+            ->method('isFeatureEnabled')
+            ->with('product_original_filenames')
+            ->willReturn(true);
+
+        self::assertEquals('filename-original-filename_-123-картинка.jpeg', $this->provider->getFileName($file));
+    }
+
+    public function getExtensionDataProvider(): array
+    {
+        return [
+            'no extension' => [
+                'extension' => null,
+            ],
+            'extension' => [
+                'extension' => 'jpeg',
+            ],
+        ];
     }
 }